Create new resume on profile creation and use that on profile builder
