+++ src/app/api/cv/[id]/route.ts
@@ -1,0 +1,93 @@
+import { NextRequest, NextResponse } from 'next/server';
+import { getToken } from 'next-auth/jwt';
+import prisma from '../../../../lib/prisma';
+
+export async function GET(req: NextRequest) {
+  console.log('Getting stored profile...');
+
+  const token = await getToken({ req });
+  if (!token && token != null) {
+    console.log('invalid token');
+    return NextResponse.json(
+      {
+        body: {
+          message: 'invalid token',
+        },
+      },
+      { status: 401 },
+    );
+  }
+
+  if (!token?.sub) {
+    console.log('No user ID provided');
+    return NextResponse.json(
+      {
+        body: {
+          message: 'no user ID provided',
+        },
+      },
+      { status: 400 },
+    );
+  }
+
+  const cvId = req.nextUrl.pathname.split('/').pop();
+  if (!cvId) {
+    console.log('No CV ID provided');
+    return NextResponse.json(
+      {
+        body: {
+          message: 'no CV ID provided',
+        },
+      },
+      { status: 400 },
+    );
+  }
+
+  console.log(`Fetching profile id: ${cvId}`);
+
+  let structuredCV;
+  try {
+    structuredCV = await prisma.structuredCVs.findUnique({
+      where: {
+        ownerId: token?.sub,
+        id: cvId,
+      },
+      select: {
+        id: true,
+        content: true,
+      },
+    });
+  } catch (err) {
+    console.log(`Fetching parsed CV: ${err}`);
+    return NextResponse.json(
+      {
+        body: {
+          message: 'internal server error',
+        },
+      },
+      { status: 500 },
+    );
+  }
+
+  if (!structuredCV || !structuredCV.content) {
+    console.log('No resume found');
+    return NextResponse.json(
+      {
+        body: {
+          message: 'no resume found',
+        },
+      },
+      { status: 404 },
+    );
+  }
+
+  return NextResponse.json(
+    {
+      body: {
+        id: structuredCV?.id,
+        profile: structuredCV?.content,
+      },
+    },
+    { status: 200 },
+  );
+}
